fix(groupby): removing density component for now #856
