+++ projects/igniteui-angular/src/lib/chips/chip.component.ts
@@ -20,7 +20,7 @@
 import { IgxRippleModule } from '../directives/ripple/ripple.directive';
 import { IgxSuffixDirective } from '../directives/suffix/suffix.directive';
 import { IgxDragDirective } from '../directives/dragdrop/dragdrop.directive';
-import { IgxDensityEnabledComponent } from '../core/density';
+import { DisplayDensity } from '../core/utils';
 
 @Component({
     selector: 'igx-chip',
@@ -36,7 +36,7 @@ import { IgxDensityEnabledComponent } from '../core/density';
         `
     ]
 })
-export class IgxChipComponent extends IgxDensityEnabledComponent {
+export class IgxChipComponent {
 
     @Input()
     public id;
@@ -47,8 +47,36 @@ export class IgxChipComponent extends IgxDensityEnabledComponent {
     @Input()
     public removable = true;
 
-    @HostBinding('class.igx-chip')
-    public cssClass = 'igx-chip';
+    @HostBinding('attr.class')
+    get hostClass(): string {
+        switch (this._displayDensity) {
+            case DisplayDensity.cosy:
+                return 'igx-chip--cosy';
+            case DisplayDensity.compact:
+                return 'igx-chip--compact';
+            default:
+                return 'igx-chip';
+        }
+    }
+
+    @Input()
+    public get displayDensity(): DisplayDensity | string {
+        return this._displayDensity;
+    }
+
+    public set displayDensity(val: DisplayDensity | string) {
+        switch (val) {
+            case 'compact':
+                this._displayDensity = DisplayDensity.compact;
+                break;
+            case 'cosy':
+                this._displayDensity = DisplayDensity.cosy;
+                break;
+            case 'comfortable':
+            default:
+                this._displayDensity = DisplayDensity.comfortable;
+        }
+    }
 
     @Input()
     public set color(newColor) {
@@ -86,13 +114,9 @@ export class IgxChipComponent extends IgxDensityEnabledComponent {
 
     public areaMovingPerforming = false;
 
-    constructor(public cdr: ChangeDetectorRef, public elementRef: ElementRef) {
-        super();
-    }
+    private _displayDensity = DisplayDensity.comfortable;
 
-    protected get hostClassPrefix() {
-        return 'igx-chip';
-    }
+    constructor(public cdr: ChangeDetectorRef, public elementRef: ElementRef) { }
 
     public onChipRemove() {
         this.onRemove.emit({
