+++ projects/igniteui-angular/src/lib/grid/grid-summary.component.ts
@@ -2,12 +2,12 @@ import {
     AfterContentInit, ChangeDetectionStrategy, ChangeDetectorRef,
     Component, DoCheck, HostBinding, HostListener, Input, OnDestroy, OnInit
 } from '@angular/core';
+import { DisplayDensity } from '../core/utils';
 import { DataType } from '../data-operations/data-util';
 import { IgxGridAPIService } from './api.service';
 import { IgxColumnComponent } from './column.component';
 import { IgxDateSummaryOperand, IgxNumberSummaryOperand, IgxSummaryOperand, IgxSummaryResult } from './grid-summary';
 import { autoWire, IGridBus } from './grid.common';
-import { IgxDensityEnabledComponent } from '../core/density';
 
 @Component({
     changeDetection: ChangeDetectionStrategy.OnPush,
@@ -15,8 +15,7 @@ import { IgxDensityEnabledComponent } from '../core/density';
     selector: 'igx-grid-summary',
     templateUrl: './grid-summary.component.html'
 })
-export class IgxGridSummaryComponent extends IgxDensityEnabledComponent
-     implements IGridBus, OnInit, DoCheck, AfterContentInit {
+export class IgxGridSummaryComponent implements IGridBus, OnInit, DoCheck, AfterContentInit {
 
     fieldName: string;
 
@@ -28,6 +27,19 @@ export class IgxGridSummaryComponent extends IgxDensityEnabledComponent
 
     get dataType(): DataType {
         return this.column.dataType;
+    }
+
+    @HostBinding('attr.class')
+    get defaultClass(): string {
+        switch (this.displayDensity) {
+            case DisplayDensity.compact:
+                return 'igx-grid-summary--compact';
+            case DisplayDensity.cosy:
+                return 'igx-grid-summary--cosy';
+            case DisplayDensity.comfortable:
+            default:
+                return 'igx-grid-summary';
+        }
     }
 
     @HostBinding('class.igx-grid-summary--fw')
@@ -66,10 +78,9 @@ export class IgxGridSummaryComponent extends IgxDensityEnabledComponent
     private hiddenItemClass = 'igx-grid-summary__item--inactive';
     private summaryResultClass = 'igx-grid-summary-item__result--left-align';
     private numberSummaryResultClass = 'igx-grid-summary-item__result';
+    private displayDensity: DisplayDensity | string;
 
-    constructor(public gridAPI: IgxGridAPIService, public cdr: ChangeDetectorRef) {
-        super();
-    }
+    constructor(public gridAPI: IgxGridAPIService, public cdr: ChangeDetectorRef) { }
 
     @autoWire(true)
     public ngOnInit() {
@@ -81,7 +92,7 @@ export class IgxGridSummaryComponent extends IgxDensityEnabledComponent
 
     ngAfterContentInit() {
         this.displayDensity = this.gridAPI.get(this.gridID).displayDensity;
-        this.summaryItemHeight = this.defaultRowHeight;
+        this.summaryItemHeight = this.gridAPI.get(this.gridID).defaultRowHeight;
     }
 
     get resolveSummaries(): any[] {
