--- projects/igniteui-angular/src/lib/grid/groupby-drop-area.component.html
@@ -1,3 +1,0 @@
-<igx-icon fontSet="material" name="group_work" class="igx-drop-area__icon"></igx-icon>
-
-<span class="igx-drop-area__text">Drag a column header and drop it here to group by that column</span>
